Support merging consensus ballots
This commit implements a merging strategy for consensus ballots, making
sure that speaker order, ghosts, as well as declared winners and speaker
scores, are identical, lest an error be raised.

Now, rather than aborting the merge when an error is detected, the
error(s) are applied to the new ballot, blanking the fields that were
errored. It also shows more general errors of the detected divergences,
rather than pointing out where it thinks the error was made. That wasn't
needed anymore as the specific fields are now obvious.

If the speaker order is inconsistent, should also throw out the scores
since they may have been assigned to the wrong speaker.